Extracting out webdriver proxy / Adding thumbnail tests to CI
