+++ superset/tasks/thumbnails.py
@@ -18,17 +18,16 @@
 """Utility functions used across Superset"""
 
 import logging
-from typing import Optional, Tuple
+from typing import Optional
 
 from flask import current_app
 
 from superset import app, security_manager, thumbnail_cache
 from superset.extensions import celery_app
 from superset.utils.screenshots import ChartScreenshot, DashboardScreenshot
+from superset.utils.webdriver import WindowSize
 
 logger = logging.getLogger(__name__)
-
-WindowSize = Tuple[int, int]
 
 
 @celery_app.task(name="cache_chart_thumbnail", soft_time_limit=300)
